+++ services/email/email.js
@@ -1,14 +1,7 @@
 const broker = require('../init')
-
-const mailgun = require('mailgun-js')({
-  apiKey: process.env.MAILGUN_API_KEY || 'secret-key-here',
-  domain: process.env.DOMAIN || 'www.gmail.com',
-  host: process.env.MAILGUN_HOST || 'api.mailgun.net'
-})
-const sender = process.env.SENDER_EMAIL || 'username username@gmail.com'
 const auth = process.env.AUTH || 'secret-key-here'
-const sgMail = require('@sendgrid/mail')
-sgMail.setApiKey(process.env.SENGRID_API_KEY)
+const mailGun = require('./methods/mailgun')
+const senGrid = require('./methods/sengrid')
 
 broker.createService({
   settings: {
@@ -17,33 +10,19 @@ broker.createService({
   name: 'email',
   actions: {
     async send (ctx) {
+      let response = ''
       if (ctx.params.auth !== auth) {
         ctx.meta.$statusCode = 401
         return { error: 'Error: auth key invalid' }
       }
-      let res = ''
+
       try {
         if (Math.floor(Math.random() * 2) % 2) {
-          const data = {
-            from: sender,
-            to: ctx.params.to,
-            subject: ctx.params.subject,
-            text: ctx.params.text,
-            html: ctx.params.html,
-            'recipient-variables': '{}'
-          }
-          res = await mailgun.messages().send(data)
+          response = await senGrid.sengrid(ctx)
         } else {
-          const data = {
-            from: sender,
-            to: ctx.params.to.split(', '),
-            subject: ctx.params.subject,
-            text: ctx.params.text,
-            html: ctx.params.html
-          }
-          res = await sgMail.sendMultiple(data)
+          response = await mailGun.mailgun(ctx)
         }
-        return res
+        return response
       } catch (err) {
         ctx.meta.$statusCode = err.statusCode
         return { error: err.toString() }
